requirements.txt
beautifulsoup4
cld2-cffi
flask
ftfy
langid
requests
waitress
